+++ changelog.md
@@ -7,6 +7,8 @@
     - support for *input* patter using globs
 - formatter built on top of *Typescript formatter*
 - *formatter*
+    - supports formating of files
+    - supports formating of provided content string
     - supports reformating of *import* statements
     - supports reordering of *import* statements
     - supports reformating of *decorator* declarations